Fixed type mismatch
git-svn-id: http://usg.lofar.org/svn/code/trunk/src/DAL@2906 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ implement/Coordinate.h
@@ -115,10 +115,10 @@ namespace DAL   // Namespace DAL -- begin
       Coordinate ();
       //! Argumented constructor
       Coordinate (Coordinate::Type const &coordinateType,
-                  double const &nofAxes);
+                  unsigned int const &nofAxes);
       //! Argumented constructor
       Coordinate (Coordinate::Type const &coordinateType,
-                  double const &nofAxes,
+                  unsigned int const &nofAxes,
                   std::vector<std::string> const &axisNames,
                   std::vector<std::string> const &axisUnits,
                   std::vector<double> const &refValue,
